--- src/Revit.TestRunner.SampleTestProject/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle( "Revit.TestRunner.SampleTestProject" )]
-
-[assembly: ComVisible( false )]
-[assembly: Guid( "02702b89-16ea-4cc1-85f4-6cd7e2790dec" )]
-
