--- src/Revit.TestRunner.SampleTestProject/bin/Debug/netstandard2.0/Revit.TestRunner.SampleTestProject.deps.json
@@ -1,94 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETStandard,Version=v2.0/",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETStandard,Version=v2.0": {},
-    ".NETStandard,Version=v2.0/": {
-      "Revit.TestRunner.SampleTestProject/1.0.0": {
-        "dependencies": {
-          "NETStandard.Library": "2.0.3",
-          "NUnit": "3.13.2",
-          "RevitAPI": "22.0.0.0",
-          "RevitAPIUI": "22.0.0.0"
-        },
-        "runtime": {
-          "Revit.TestRunner.SampleTestProject.dll": {}
-        }
-      },
-      "Microsoft.NETCore.Platforms/1.1.0": {},
-      "NETStandard.Library/2.0.3": {
-        "dependencies": {
-          "Microsoft.NETCore.Platforms": "1.1.0"
-        }
-      },
-      "NUnit/3.13.2": {
-        "dependencies": {
-          "NETStandard.Library": "2.0.3"
-        },
-        "runtime": {
-          "lib/netstandard2.0/nunit.framework.dll": {
-            "assemblyVersion": "3.13.2.0",
-            "fileVersion": "3.13.2.0"
-          }
-        }
-      },
-      "RevitAPI/22.0.0.0": {
-        "runtime": {
-          "RevitAPI.dll": {
-            "assemblyVersion": "22.0.0.0",
-            "fileVersion": "22.0.2.392"
-          }
-        }
-      },
-      "RevitAPIUI/22.0.0.0": {
-        "runtime": {
-          "RevitAPIUI.dll": {
-            "assemblyVersion": "22.0.0.0",
-            "fileVersion": "22.0.2.392"
-          }
-        }
-      }
-    }
-  },
-  "libraries": {
-    "Revit.TestRunner.SampleTestProject/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    },
-    "Microsoft.NETCore.Platforms/1.1.0": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-kz0PEW2lhqygehI/d6XsPCQzD7ff7gUJaVGPVETX611eadGsA3A877GdSlU0LRVMCTH/+P3o2iDTak+S08V2+A==",
-      "path": "microsoft.netcore.platforms/1.1.0",
-      "hashPath": "microsoft.netcore.platforms.1.1.0.nupkg.sha512"
-    },
-    "NETStandard.Library/2.0.3": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-st47PosZSHrjECdjeIzZQbzivYBJFv6P2nv4cj2ypdI204DO+vZ7l5raGMiX4eXMJ53RfOIg+/s4DHVZ54Nu2A==",
-      "path": "netstandard.library/2.0.3",
-      "hashPath": "netstandard.library.2.0.3.nupkg.sha512"
-    },
-    "NUnit/3.13.2": {
-      "type": "package",
-      "serviceable": true,
-      "sha512": "sha512-u+fz/lXyR4vlamySNAEMrXvh+GhAQiB6/aVZtU5WjivR5zF26Ui0tfteDtWqT90k9D8y6g8rFKYQC97Z7d195w==",
-      "path": "nunit/3.13.2",
-      "hashPath": "nunit.3.13.2.nupkg.sha512"
-    },
-    "RevitAPI/22.0.0.0": {
-      "type": "reference",
-      "serviceable": false,
-      "sha512": ""
-    },
-    "RevitAPIUI/22.0.0.0": {
-      "type": "reference",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
