+++ src/Revit.TestRunner.SampleTestProject2/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle( "Revit.TestRunner.SampleTestProject2" )]
+[assembly: ComVisible( false )]
+
+[assembly: Guid( "c797887e-e708-4f53-b1b8-b1d28fc24c8b" )]
