+++ src/Revit.TestRunner.SampleTestProject2/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="NUnit" version="3.13.2" targetFramework="net48" />
+</packages>
